# RequiredNamedArgumentsAnalyzers: fixes and optimizations