With ImmutabilityAnalyzer as reference

## Before
```
D2L.CodeStyle.Analyzers.Immutability.ImmutabilityAnalyzer = 11.690 s
D2L.CodeStyle.Analyzers.Language.RequireNamedArgumentsAnalyzer = 10.423 s
```

## After
```
D2L.CodeStyle.Analyzers.Immutability.ImmutabilityAnalyzer = 11.804 s
D2L.CodeStyle.Analyzers.Language.RequireNamedArgumentsAnalyzer = 9.463 s
```